+++ CeilingFan/usermessage.py
@@ -1,0 +1,3 @@
+# Function to return final statement from action selected
+def return_success_msg(user_fan_name, action_type):
+    print(f"Fan named {user_fan_name} has been successfully {action_type}.")
